+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### BUG FIXES
 
+- `[consensus]` prevent ingest response panic during shutdown
+  ([\#6065](https://github.com/cometbft/cometbft/pull/6065))
 - `[lp2p]` fix flaky MsgBytesFilter test by avoiding broadcast race
   ([\#6053](https://github.com/cometbft/cometbft/pull/6053))
 - `[spec]` fix the inductive invariant `spec/light-client/accountability`